reorg files and address aditi feedback
