+++ aries-site/src/examples/components/namevaluelist/DoDonts/NameValueListTextGoodExample.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { Box, NameValueList, NameValuePair, Text } from 'grommet';
-import { data } from './data';
+import { data } from '../data';
 
-export const NameValueListTextBadExample = () => (
+export const NameValueListTextGoodExample = () => (
   <Box pad="small">
-    <NameValueList>
+    <NameValueList nameProps={{ width: 'xxsmall' }}>
       {Object.entries(data).map(([name, value]) => (
-        <NameValuePair key={name} name={<Text size="xxlarge">{name}</Text>}>
+        <NameValuePair key={name} name={<Text size="xsmall">{name}</Text>}>
           {value}
         </NameValuePair>
       ))}
