+++ aries-site/src/examples/components/namevaluelist/DoDonts/NameValueListWeightBad.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Box, NameValueList, NameValuePair, Text } from 'grommet';
-import { serverData } from './data';
+import { serverData } from '../data';
 
 export const NameValueListWeightBadExample = () => (
   <Box pad="small">
