+++ aries-site/src/examples/components/namevaluelist/DoDonts/NameValueListWeightGood.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Box, NameValueList, NameValuePair, Text } from 'grommet';
-import { serverData } from './data';
+import { serverData } from '../data';
 
 export const NameValueListWeightGoodExample = () => (
   <Box pad="small">
