+++ aries-site/src/examples/components/namevaluelist/Examples/NameValueListSimpleExample.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Box, NameValueList, NameValuePair } from 'grommet';
-import { fontData } from './data';
+import { simpleData } from '../data';
 
-export const NameValueListFontWeightExample = () => (
+export const NameValueListSimpleExample = () => (
   <Box pad="small">
     <NameValueList>
-      {Object.entries(fontData).map(([name, value]) => (
+      {Object.entries(simpleData).map(([name, value]) => (
         <NameValuePair key={name} name={name}>
           {value}
         </NameValuePair>
