+++ aries-site/src/examples/components/namevaluelist/Previews/NameValueListAnatomyPreview.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Box, NameValueList, NameValuePair } from 'grommet';
-import { languageData } from './data';
+import { anatomyData } from '../data';
 
-export const NameValueListBoldExampleGood = () => (
+export const NameValueListAnatomyPreview = () => (
   <Box pad="small">
     <NameValueList>
-      {Object.entries(languageData).map(([name, value]) => (
+      {Object.entries(anatomyData).map(([name, value]) => (
         <NameValuePair key={name} name={name}>
           {value}
         </NameValuePair>
