+++ aries-site/src/examples/components/namevaluelist/Previews/NameValueListFontWeightPreview.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Box, NameValueList, NameValuePair } from 'grommet';
-import { simpleData } from './data';
+import { fontData } from '../data';
 
-export const NameValueListPairPropsExample = () => (
+export const NameValueListFontWeightPreview = () => (
   <Box pad="small">
-    <NameValueList pairProps={{ direction: 'column' }}>
-      {Object.entries(simpleData).map(([name, value]) => (
+    <NameValueList>
+      {Object.entries(fontData).map(([name, value]) => (
         <NameValuePair key={name} name={name}>
           {value}
         </NameValuePair>
