+++ aries-site/src/examples/components/namevaluelist/index.js
@@ -1,20 +1,5 @@
-export * from './NameValueListAlignmentPreview';
+export * from './DoDonts';
+export * from './Examples';
 export * from './NameValueListAnatomy';
-export * from './NameValueListAnatomyExample';
-export * from './NameValueListBoldExampleBad';
-export * from './NameValueListBoldExampleGood';
-export * from './NameValueListDefaultExample';
-export * from './NameValueListFontWeightExample';
-export * from './NameValueListFontWeightSecondExample';
-export * from './NameValueListGridExample';
-export * from './NameValueListHeadingExample';
-export * from './NameValueListIconNamePreview';
-export * from './NameValueListIconValuePreview';
-export * from './NameValueListPairPropsExample';
-export * from './NameValueListSecondHeadingExample';
 export * from './NameValueListScaleTable';
-export * from './NameValueListSimpleExample';
-export * from './NameValueListTextBadExample';
-export * from './NameValueListTextGoodExample';
-export * from './NameValueListWeightGood';
-export * from './NameValueListWeightBad';
+export * from './Previews';
