+++ aries-site/src/pages/components/namevaluelist.mdx
@@ -3,25 +3,23 @@ import { ColorSwatch, ExampleImagePreview } from '../../components';
 import {
   NameValueListAlignmentPreview,
   NameValueListAnatomy,
-  NameValueListAnatomyExample,
+  NameValueListAnatomyPreview,
   NameValueListBoldExampleBad,
   NameValueListBoldExampleGood,
   NameValueListDefaultExample,
-  NameValueListFontWeightExample,
-  NameValueListFontWeightSecondExample,
+  NameValueListFontWeightPreview,
+  NameValueListFontWeightSecondPreview,
   NameValueListGridExample,
-  NameValueListHeadingExample,
+  NameValueListHeadingPreview,
   NameValueListIconNamePreview,
   NameValueListIconValuePreview,
   NameValueListPairPropsExample,
-  NameValueListSecondHeadingExample,
   NameValueListScaleTable,
   NameValueListSimpleExample,
   NameValueListTextBadExample,
   NameValueListTextGoodExample,
   NameValueListWeightBadExample,
-  NameValueListWeightGoodExample,
-  
+  NameValueListWeightGoodExample, 
 } from '../../examples';
 
 <Example
@@ -59,7 +57,7 @@ A NameValuePair is a set of two related elements: a name, which identifies a dat
    Value is the description of data or a pointer to the location of that data. The format of the value can vary, depending on the data set. Content can include: single line description, multi line description, anchors, array or list of items.
 
   <Example caption="Example: value can include different types of content" pad="none" height={{ min: 'small' }}>
-   <NameValueListAnatomyExample />
+   <NameValueListAnatomyPreview />
   </Example>
 
 3. **Supporting icons/visuals (optional)**
@@ -134,12 +132,8 @@ to imply hierarchy: [order](#order), [headings](#headings), [scale](#scale),
 - For accessibility reasons, always ensure semantically correct heading levels are applied. See [best practices for applying headings](/foundation/typography#semantic-usage-of-heading-levels).
 
  <Example content="Example: An h2 small heading is applied to a vertical name-value list" pad="none" height={{ min: 'small' }}>
-  <NameValueListSecondHeadingExample />
+  <NameValueListHeadingPreview />
  </Example>
-
-  <Example content="Example: An h2 small heading is applied to a stacked name-value list" pad="none" height={{ min: 'small' }}>
-    <NameValueListHeadingExample />
-  </Example>
 
 ### Scale
 
@@ -176,10 +170,10 @@ Making the font bolder or thinner is another way to create hierarchy that’s ea
 - Don’t use a bold font-weight for any long format content as this can be very hard to read and overwhelm the user.
 
   <Example caption="Example: A two column Name-Value List where the names are in a bold font-weight" pad="none" height="none">
-    <NameValueListFontWeightExample />
+    <NameValueListFontWeightPreview />
   </Example>
   <Example caption="Example: A multi column Name-Value List where the values are in a bold font-weight." pad="small" height="none" width="large">
-    <NameValueListFontWeightSecondExample />
+    <NameValueListFontWeightSecondPreview />
   </Example>  
 
 <BestPracticeGroup>
@@ -228,9 +222,9 @@ All types of NameValueLists will change to a single column NameValueList on smal
 By default, both the name and the value are left aligned. This can be changed through `nameProps` and `valueProps` by setting [align](https://v2.grommet.io/namevaluelist?theme=hpe#nameProps) to the desired value, such as `end`.
 No matter what alignment is chosen, always opt for a consistent presentation across a single NameValueList or NameValueLists on a single page.
 
-<Example caption="Left aligned example of a vertical name-value list:" pad="small" height="none">
-  <NameValueListAlignmentPreview />
-</Example>  
+  <Example caption="Left aligned example of a vertical name-value list:" pad="small" height="none">
+   <NameValueListAlignmentPreview />
+  </Example>  
 
 ### Controlling the width of the name and value
 The width of name is set to "small" by default to accommodate a reasonable width for common data names. However, if all names are short, it may be appropriate to adjust the name width to something smaller. This can be easily changed through `nameProps`. For example `nameProps={{width: 'xsmall'}}`.
